warehouse_manager/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
asttokens==2.4.1
auto-py-to-exe==2.43.3
Babel==2.14.0
bottle==0.12.25
bottle-websocket==0.2.9
build==1.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
Eel==0.16.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==22.0.0
humanfriendly==10.0
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mkl==2021.4.0
mod-wsgi @ file:///C:/Users/Administrador/Downloads/mod_wsgi-4.9.2-cp311-cp311-win_amd64.whl#sha256=f5cd8beff2f4730909eb351086964b8ffcf111226eee3ca5576e95e42d5d4c8d
mpmath==1.3.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
openpyxl==3.1.2
overpy==0.7
packaging==24.0
pandas==2.2.1
parso==0.8.3
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
pure-eval==0.2.2
pycocotools==2.0.7
pycparser==2.22
Pygments==2.17.2
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyMySQL==1.1.0
pyodbc==5.1.0
pyparsing==3.1.2
pypng==0.20220715.0
pyproject_hooks==1.0.0
pypyodbc==1.3.6
pyreadline3==3.4.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
qrcode==7.4.2
requests==2.32.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tkcalendar==1.6.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
wcwidth==0.2.13
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
